Refactor package structure and fix import errors in GameRoomController, GameRoomObservable, GameRoomObserver, and MusicPlayer
