+++ core/src/no/ntnu/game/Controllers/GameRoomObservable.java
@@ -1,0 +1,14 @@
+package no.ntnu.game.Controllers;
+
+import no.ntnu.game.firestore.GameRoom.GameStatus;
+
+public interface GameRoomObservable {
+
+    void setGameStatus(GameStatus gameStatus);
+
+    GameStatus getGameStatus();
+
+    void addObserver(GameRoomObserver observer);
+
+    void removeObserver(GameRoomObserver observer);
+}
