+++ core/src/no/ntnu/game/Controllers/GameRoomObserver.java
@@ -1,0 +1,7 @@
+package no.ntnu.game.Controllers;
+
+import no.ntnu.game.firestore.GameRoom;
+
+public interface GameRoomObserver {
+    void gameStatusChanged(GameRoom gameRoom);
+}
