--- core/src/no/ntnu/game/firestore/GameRoomObservable.java
@@ -1,14 +1,0 @@
-package no.ntnu.game.firestore;
-
-import no.ntnu.game.firestore.GameRoom.GameStatus;
-
-public interface GameRoomObservable {
-
-    void setGameStatus(GameStatus gameStatus);
-
-    GameStatus getGameStatus();
-
-    void addObserver(GameRoomObserver observer);
-
-    void removeObserver(GameRoomObserver observer);
-}
