--- core/src/no/ntnu/game/firestore/GameRoomObserver.java
@@ -1,5 +1,0 @@
-package no.ntnu.game.firestore;
-
-public interface GameRoomObserver {
-    void gameStatusChanged(GameRoom gameRoom);
-}
